main 브랜치로 병합

See merge request 23_HF141/studious-backend!87

+++ src/main/java/nerds/studiousTestProject/review/service/ReviewService.java
@@ -2,27 +2,29 @@ package nerds.studiousTestProject.review.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import nerds.studiousTestProject.common.exception.BadRequestException;
 import nerds.studiousTestProject.common.exception.NotFoundException;
-import nerds.studiousTestProject.common.service.StorageService;
-import nerds.studiousTestProject.common.service.TokenService;
+import nerds.studiousTestProject.common.service.StorageProvider;
 import nerds.studiousTestProject.hashtag.entity.HashtagName;
 import nerds.studiousTestProject.hashtag.entity.HashtagRecord;
 import nerds.studiousTestProject.hashtag.repository.HashtagRecordRepository;
 import nerds.studiousTestProject.member.entity.member.Member;
+import nerds.studiousTestProject.member.repository.MemberRepository;
 import nerds.studiousTestProject.photo.entity.SubPhoto;
 import nerds.studiousTestProject.photo.entity.SubPhotoType;
 import nerds.studiousTestProject.photo.repository.SubPhotoRepository;
 import nerds.studiousTestProject.reservation.entity.ReservationRecord;
+import nerds.studiousTestProject.reservation.entity.ReservationStatus;
 import nerds.studiousTestProject.reservation.repository.ReservationRecordRepository;
-import nerds.studiousTestProject.review.dto.available.response.AvailableReviewResponse;
-import nerds.studiousTestProject.review.dto.find.response.TotalGradeInfo;
-import nerds.studiousTestProject.review.dto.modify.request.ModifyReviewRequest;
-import nerds.studiousTestProject.review.dto.register.request.RegisterReviewRequest;
 import nerds.studiousTestProject.review.dto.available.response.AvailableReviewInfo;
+import nerds.studiousTestProject.review.dto.available.response.AvailableReviewResponse;
 import nerds.studiousTestProject.review.dto.delete.response.DeleteReviewResponse;
 import nerds.studiousTestProject.review.dto.find.response.FindReviewInfo;
 import nerds.studiousTestProject.review.dto.find.response.FindReviewSortedResponse;
+import nerds.studiousTestProject.review.dto.find.response.TotalGradeInfo;
+import nerds.studiousTestProject.review.dto.modify.request.ModifyReviewRequest;
 import nerds.studiousTestProject.review.dto.modify.response.ModifyReviewResponse;
+import nerds.studiousTestProject.review.dto.register.request.RegisterReviewRequest;
 import nerds.studiousTestProject.review.dto.register.response.RegisterReviewResponse;
 import nerds.studiousTestProject.review.dto.written.response.WrittenReviewInfo;
 import nerds.studiousTestProject.review.dto.written.response.WrittenReviewResponse;
@@ -38,13 +40,18 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static nerds.studiousTestProject.common.exception.ErrorCode.NOT_FOUND_RESERVATION_RECORD;
-import static nerds.studiousTestProject.common.exception.ErrorCode.NOT_FOUND_REVIEW;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.EXPIRED_VALID_DATE;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.INVALID_RESERVATION_STATUS;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.INVALID_WRITE_REVIEW_TIME;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.NOT_FOUND_RESERVATION_RECORD;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.NOT_FOUND_REVIEW;
+import static nerds.studiousTestProject.common.exception.errorcode.ErrorCode.NOT_FOUND_USER;
 
 
 @RequiredArgsConstructor
@@ -52,17 +59,21 @@ import static nerds.studiousTestProject.common.exception.ErrorCode.NOT_FOUND_REV
 @Service
 @Transactional(readOnly = true)
 public class ReviewService {
+
     private final HashtagRecordRepository hashtagRecordRepository;
+    private final MemberRepository memberRepository;
     private final ReservationRecordRepository reservationRecordRepository;
     private final ReviewRepository reviewRepository;
-    private final StorageService storageService;
+    private final StorageProvider storageProvider;
     private final SubPhotoRepository subPhotoRepository;
-    private final TokenService tokenService;
     public final Double GRADE_COUNT = 3.0;
-
 
     @Transactional
     public RegisterReviewResponse register(RegisterReviewRequest registerReviewRequest, List<MultipartFile> files){
+        ReservationRecord reservationRecord = findReservationRecordById(registerReviewRequest.getReservationId());
+        validateReservationStatus(reservationRecord);
+        validateAvailableReviewTime(reservationRecord);
+
         Grade grade = RegisterReviewRequest.toGrade(registerReviewRequest);
         grade.updateTotal(getTotal(grade.getCleanliness(), grade.getDeafening(), grade.getFixturesStatus()));
 
@@ -74,7 +85,6 @@ public class ReviewService {
                 .build();
         reviewRepository.save(review);
 
-        ReservationRecord reservationRecord = findReservationRecordById(registerReviewRequest.getReservationId());
         reservationRecord.addReview(review);
 
         List<String> hashtags = registerReviewRequest.getHashtags();
@@ -100,12 +110,10 @@ public class ReviewService {
         Review review = findReviewById(reviewId);
 
         Grade grade = review.getGrade();
-        grade.updateGrade(modifyReviewRequest.getCleanliness(),
+        grade.update(modifyReviewRequest.getCleanliness(),
                 modifyReviewRequest.getDeafening(),
                 modifyReviewRequest.getFixtureStatus(),
                 getTotal(grade.getCleanliness(), grade.getDeafening(), grade.getFixturesStatus()));
-
-        // 추천 여부 수정 만들어야 함
 
         review.getHashtagRecords().removeAll(review.getHashtagRecords());
         deleteAllHashtagRecordByReviewId(reviewId);
@@ -118,19 +126,23 @@ public class ReviewService {
             review.addHashtagRecord(hashtagRecord);
         }
 
-        // 사진은 리뷰id를 통해 삭제하고, 다시 받아온 url로 저장을 한다.
         deleteAllPhotos(reviewId);
         saveSubPhotos(review, files);
 
         review.updateDetail(modifyReviewRequest.getDetail());
+        review.updateIsRecommended(modifyReviewRequest.getIsRecommend());
 
         return ModifyReviewResponse.builder().reviewId(reviewId).modifiedAt(LocalDate.now()).build();
     }
 
     @Transactional
     public DeleteReviewResponse deleteReview(Long reviewId) {
-        Review review = findReviewById(reviewId);
         ReservationRecord reservationRecord = findReservationRecordByReviewId(reviewId);
+        if (LocalDate.now().isBefore(reservationRecord.getDate().plusDays(7))) {
+            throw new BadRequestException(EXPIRED_VALID_DATE);
+        }
+
+        Review review = findReviewById(reviewId);
         review.getHashtagRecords().removeAll(review.getHashtagRecords());
 
         deleteAllHashtagRecordByReviewId(reviewId);
@@ -171,8 +183,8 @@ public class ReviewService {
                 .build();
     }
 
-    public AvailableReviewResponse findAvailableReviews(String accessToken, Pageable pageable) {
-        Page<ReservationRecord> reservationRecords = getReservationRecords(accessToken, pageable);
+    public AvailableReviewResponse findAvailableReviews(Long memberId, Pageable pageable) {
+        Page<ReservationRecord> reservationRecords = getReservationRecords(memberId, pageable);
 
         return AvailableReviewResponse.builder()
                 .totalPage(reservationRecords.getTotalPages())
@@ -181,8 +193,8 @@ public class ReviewService {
                 .build();
     }
 
-    public WrittenReviewResponse findWrittenReviews(String accessToken, LocalDate startDate, LocalDate endDate, Pageable pageable) {
-        Page<ReservationRecord> reservationRecords = getReservationRecords(accessToken, pageable);
+    public WrittenReviewResponse findWrittenReviews(Long memberId, LocalDate startDate, LocalDate endDate, Pageable pageable) {
+        Page<ReservationRecord> reservationRecords = getReservationRecords(memberId, pageable);
 
         return WrittenReviewResponse.builder()
                 .totalPage(reservationRecords.getTotalPages())
@@ -205,6 +217,10 @@ public class ReviewService {
         List<Review> reviewList = getAllReviews(studycafeId);
         Integer count = 0, sum = 0;
 
+        if (reviewList.isEmpty()) {
+            return 0;
+        }
+
         for (Review review : reviewList){
             sum += review.getGrade().getCleanliness();
             count++;
@@ -216,6 +232,10 @@ public class ReviewService {
     public Integer getAvgDeafening(Long studycafeId) {
         List<Review> reviewList = getAllReviews(studycafeId);
         Integer count = 0, sum = 0;
+
+        if (reviewList.isEmpty()) {
+            return 0;
+        }
 
         for (Review review : reviewList){
             sum += review.getGrade().getDeafening();
@@ -229,6 +249,10 @@ public class ReviewService {
         List<Review> reviewList = getAllReviews(studycafeId);
         Integer count = 0, sum = 0;
 
+        if (reviewList.isEmpty()) {
+            return 0;
+        }
+
         for (Review review : reviewList){
             sum += review.getGrade().getFixturesStatus();
             count++;
@@ -240,6 +264,10 @@ public class ReviewService {
     public Integer getAvgRecommendation(Long studycafeId) {
         List<Review> reviewList = getAllReviews(studycafeId);
         Integer recommend = 0, count = 0;
+
+        if (reviewList.isEmpty()) {
+            return 0;
+        }
 
         for (Review review : reviewList){
             if(review.getIsRecommended()){
@@ -259,6 +287,10 @@ public class ReviewService {
         Integer count = 0;
         Double sum = 0.0;
 
+        if (reviewList.isEmpty()) {
+            return 0.0;
+        }
+
         for (Review review : reviewList){
             sum += review.getGrade().getTotal();
             count++;
@@ -269,11 +301,11 @@ public class ReviewService {
     private void saveSubPhotos(Review review, List<MultipartFile> files) {
         List<SubPhoto> photoList = new ArrayList<>();
 
-        String representPhoto = storageService.uploadFile(files.get(0));
+        String representPhoto = storageProvider.uploadFile(files.get(0));
         review.addPhoto(representPhoto);
 
         for (int i = 2; i < files.size(); i++) {
-            String uploadFile = storageService.uploadFile(files.get(i));
+            String uploadFile = storageProvider.uploadFile(files.get(i));
             photoList.add(SubPhoto.builder().review(review).type(SubPhotoType.REVIEW).path(uploadFile).build());
         }
         saveAllPhotos(photoList);
@@ -283,7 +315,7 @@ public class ReviewService {
         List<String> reviewPhotos = findAllReviewPhotos(reviewId);
 
         for (String photoUrl : reviewPhotos) {
-            storageService.deleteFile(photoUrl);
+            storageProvider.deleteFile(photoUrl);
         }
         removeAllReviewPhotos(reviewId);
     }
@@ -297,7 +329,7 @@ public class ReviewService {
         return  reservationRecordList.stream()
                 .filter(reservationRecord -> reservationRecord.getReview() == null &&
                         !reservationRecord.getDate().plusDays(7).isBefore(LocalDate.now()))
-                .map(reservationRecord -> AvailableReviewInfo.of(reservationRecord))
+                .map(reservationRecord -> AvailableReviewInfo.from(reservationRecord))
                 .sorted(Comparator.comparing(AvailableReviewInfo::getDate).reversed())
                 .collect(Collectors.toList());
     }
@@ -305,7 +337,7 @@ public class ReviewService {
     /**
      * 리뷰 작성한 내역을 조회하는 메소드
      */
-    public List<WrittenReviewInfo> getWrittenReviews(Page<ReservationRecord> reservationRecords, LocalDate startDate, LocalDate endDate) {
+    private List<WrittenReviewInfo> getWrittenReviews(Page<ReservationRecord> reservationRecords, LocalDate startDate, LocalDate endDate) {
         List<ReservationRecord> reservationRecordList = reservationRecords.getContent();
 
         return reservationRecordList.stream()
@@ -316,16 +348,13 @@ public class ReviewService {
                 .collect(Collectors.toList());
     }
 
-    private List<ReservationRecord> getAllReservation(Long studycafeId){
-        return findAllReservationRecordByStudycafeId(studycafeId);
+    private Page<ReservationRecord> getReservationRecords(Long memberId, Pageable pageable) {
+        Member member = memberRepository.findById(memberId).orElseThrow(
+                () -> new NotFoundException(NOT_FOUND_USER));
+        return findAllReservationRecordByMember(member, pageable);
     }
 
-    private Page<ReservationRecord> getReservationRecords(String accessToken, Pageable pageable) {
-        Member member = tokenService.getMemberFromAccessToken(accessToken);
-        return findAllReservationRecordByMemberId(member.getId(), pageable);
-    }
-
-    public List<FindReviewInfo> getReviewInfo(Page<Review> reviewList) {
+    private List<FindReviewInfo> getReviewInfo(Page<Review> reviewList) {
         return reviewList.stream()
                 .filter(review -> review != null && reviewList.hasContent())
                 .map(review -> FindReviewInfo.builder()
@@ -351,43 +380,17 @@ public class ReviewService {
     }
 
     private List<Review> getAllReviews(Long studycafeId) {
-        List<ReservationRecord> recordList = getAllReservation(studycafeId);
-        List<Long> reviewIds = new ArrayList<>();
-        List<Review> reviewList = new ArrayList<>();
-        for (ReservationRecord reservationRecord : recordList) {
-            reviewIds.add(reservationRecord.getReview().getId());
-        }
-
-        List<Review> reviews = reviewRepository.findAllByIdInOrderByCreatedDateDesc(reviewIds);
-
-        for (int i = 0; i < reviews.size(); i++) {
-            reviewList.add(reviews.get(i));
-        }
-        return reviewList;
+        return reviewRepository.findAllByStudycafeIdWithoutPage(studycafeId);
     }
 
     private Page<Review> getAllReviewsSorted(Long studycafeId, Pageable pageable) {
         pageable = getPageable(pageable);
-        List<ReservationRecord> recordList = getAllReservation(studycafeId);
-        return getReviewList(pageable, recordList);
+        return findAllByStudycafeId(studycafeId, pageable);
     }
 
     private Page<Review> getRoomReviewsSorted(Long studycafeId, Long roomId, Pageable pageable) {
         pageable = getPageable(pageable);
-        List<ReservationRecord> recordList = findAllReservationRecordByRoomId(roomId);
-        return getReviewList(pageable, recordList);
-    }
-
-    private Page<Review> getReviewList(Pageable pageable, List<ReservationRecord> recordList) {
-        List<Long> reviewIds = new ArrayList<>();
-
-        for (ReservationRecord reservationRecord : recordList) {
-            if(reservationRecord.getReview() != null) {
-                reviewIds.add(reservationRecord.getReview().getId());
-            }
-        }
-
-        return reviewRepository.findAllByIdIn(reviewIds, pageable);
+        return findAllByRoomId(roomId, pageable);
     }
 
     private PageRequest getPageable(Pageable pageable) {
@@ -403,6 +406,14 @@ public class ReviewService {
         return reviewRepository.findById(reviewId).orElseThrow(() -> new NotFoundException(NOT_FOUND_REVIEW));
     }
 
+    private Page<Review> findAllByStudycafeId(Long studycafeId, Pageable pageable) {
+        return reviewRepository.findAllByStudycafeId(studycafeId, pageable);
+    }
+
+    private Page<Review> findAllByRoomId(Long roomId, Pageable pageable) {
+        return reviewRepository.findAllByRoomId(roomId, pageable);
+    }
+
     private ReservationRecord findReservationRecordByReviewId(Long reviewId) {
         return reservationRecordRepository.findByReviewId(reviewId)
                 .orElseThrow(() -> new NotFoundException(NOT_FOUND_RESERVATION_RECORD));
@@ -413,16 +424,8 @@ public class ReviewService {
                 .orElseThrow(() -> new NotFoundException(NOT_FOUND_RESERVATION_RECORD));
     }
 
-    private List<ReservationRecord> findAllReservationRecordByStudycafeId(Long studycafeId) {
-        return reservationRecordRepository.findAllByStudycafeId(studycafeId);
-    }
-
-    private Page<ReservationRecord> findAllReservationRecordByMemberId(Long memberId, Pageable pageable) {
-        return reservationRecordRepository.findAllByMemberId(memberId, pageable);
-    }
-
-    private List<ReservationRecord> findAllReservationRecordByRoomId(Long roomId) {
-        return reservationRecordRepository.findAllByRoomId(roomId);
+    private Page<ReservationRecord> findAllReservationRecordByMember(Member member, Pageable pageable) {
+        return reservationRecordRepository.findAllByMember(pageable, member);
     }
 
     private void deleteAllHashtagRecordByReviewId(Long reviewId) {
@@ -446,5 +449,17 @@ public class ReviewService {
         reviewPhotos.addAll(reviewSubPhoto);
 
         return reviewPhotos;
+    }
+
+    private void validateAvailableReviewTime(ReservationRecord reservationRecord) {
+        if (LocalDate.now().isBefore(reservationRecord.getDate()) || LocalTime.now().isBefore(reservationRecord.getEndTime())) {
+            throw new BadRequestException(INVALID_WRITE_REVIEW_TIME);
+        }
+    }
+
+    private void validateReservationStatus(ReservationRecord reservationRecord) {
+        if (reservationRecord.getStatus() != ReservationStatus.CONFIRMED) {
+            throw new BadRequestException(INVALID_RESERVATION_STATUS);
+        }
     }
 }
